Fix implicit narrowings in gblib.cppm and files_rw

+++ gb/files_rw.h
@@ -5,7 +5,7 @@
 #ifndef FILES_RW_H
 #define FILES_RW_H
 
-void Fileread(int, char *, int, int);
-void Filewrite(int, char *, int, int);
+void Fileread(int, char *, size_t, int);
+void Filewrite(int, char *, size_t, int);
 
 #endif  // FILES_RW_H
